Only pass the PublicKey instead of KeyPair when the private key is not needed
The creation of the certificate itself is only about the public key that is
being certified. We only need the corresponding private key in the context
of key stores, and thus we should express that in the types.

Note that the signer key is different, because we need the private key to
sign the certificate, and the public key to verify the signature, so it makes
sense to pass the whole key pair.